Correction calcul incorrect dans l'utilisation par server

+++ include/rok4/style/Terrainrgb.h
@@ -61,7 +61,7 @@ public:
      ** \~english
      * \brief min_elevation : minimum elevation from which the color is calculated
      */
-    int min_elevation;
+    float min_elevation;
 
     /** \~french
      * \brief step : résolution verticale
